% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_int}
\alias{check_int}
\title{helper functions
Check numbers of overlapx functions}
\usage{
check_int(x, call = rlang::caller_env())
}
\arguments{
\item{x}{Input vector. numbers}

\item{call}{vector. for error report, see \code{\link[rlang:stack]{rlang::caller_env()}}}
}
\value{
logical
}
\description{
helper functions
Check numbers of overlapx functions
}
\examples{
check_int(x = c(1:10))

}
