% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{check_overlap_input}
\alias{check_overlap_input}
\title{check the input for \code{overlap}}
\usage{
check_overlap_input(x, call = rlang::caller_env())
}
\arguments{
\item{x}{Input vector. could be list, data.frame or matrix, to calculate the
overlap between groups}

\item{call}{vector. for error report, see \code{\link[rlang:stack]{rlang::caller_env()}}}
}
\value{
the input vector
}
\description{
check the input for \code{overlap}
}
\examples{
x <- list(A = 1:3, B = 2:4)
check_overlap_input(x)

}
