% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap}
\alias{overlap}
\title{overlap
generate overlap between groups
generate overlap between groups
\code{overlap} create venn plot for multiple groups}
\usage{
overlap(
  x,
  labels = c("A", "B"),
  title = "Overlap",
  quantities = "counts",
  png_file = NULL,
  height = 5,
  width = 5,
  dpi = 150,
  colors = c("#FF6D60", "#176B87", "#FF8400", "#435B66"),
  ...
)
}
\arguments{
\item{x}{Input vector. could be list, data.frame or matrix, to calculate the
overlap between groups}

\item{labels}{Input vector. The name of the groups}

\item{title}{Input vector. The title of the plot, default: Overlap}

\item{quantities}{character. could be \code{counts} or \code{percent} or both}

\item{png_file}{Input vector. NULL or path to the output png file, default
: NULL}

\item{height}{Input vector. The height of output file in inches, default: 5}

\item{width}{Input vector. The width of output file in inches, default: 5}

\item{dpi}{Input vector. The resolution of output file, default: 150}

\item{colors}{character. The colors for each group,}

\item{...}{A series of named character vectors. parameters for \code{\link[base:plot]{base::plot()}}
function}
}
\value{
A plot
}
\description{
overlap
generate overlap between groups
generate overlap between groups
\code{overlap} create venn plot for multiple groups
}
\examples{
x <- list(A = 1:10, B = 5:20)
overlap(x, labels = c("GroupA", "GroupB"))
y <- list(A = 1:10, B = 5:20, C = c(8:15, 18:25))
overlap(y)

}
\seealso{
\link[VennDiagram:venn.diagram]{VennDiagram::venn.diagram} which doing the same job
}
