% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap2}
\alias{overlap2}
\title{generate overlap between two groups
\code{overlap2} create venn plot for two groups}
\usage{
overlap2(
  n1,
  n2,
  n12,
  labels = c("A", "B"),
  title = "Overlap",
  quantities = c("counts", "percent"),
  png_file = NULL,
  height = 5,
  width = 5,
  dpi = 150,
  colors = c("#FF6D60", "#176B87"),
  ...
)
}
\arguments{
\item{n1}{Input integer. The number of components only in the first group,
must be positive integer.}

\item{n2}{Input integer. The number of components only in the second group,
must be positive integer.}

\item{n12}{Input integer. The number of components overlapped between two
groups, should be positive integer or zero.}

\item{labels}{Input vector. The name of the groups, length == 2}

\item{title}{Input vector. The title of the plot, default: Overlap}

\item{quantities}{character. could be \code{counts} or \code{percent} or both}

\item{png_file}{Input vector. NULL or path to the output png file, default
: NULL}

\item{height}{Input vector. The height of output file in inches, default: 5}

\item{width}{Input vector. The width of output file in inches, default: 5}

\item{dpi}{Input vector. The resolution of output file, default: 150}

\item{colors}{character. The colors for each group,}

\item{...}{A series of named character vectors. parameters for \code{\link[base:plot]{base::plot()}}
function}
}
\value{
A plot
}
\description{
generate overlap between two groups
\code{overlap2} create venn plot for two groups
}
\examples{
overlap2(3, 8, 1)
overlap2(12, 8, 4, name1 = "group1", name2 = "group2")

}
\seealso{
\link[VennDiagram:venn.diagram]{VennDiagram::venn.diagram} which doing the same job
}
