% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap3}
\alias{overlap3}
\title{generate overlap between three groups
\code{overlap3} create venn plot for three groups}
\usage{
overlap3(
  n1,
  n2,
  n3,
  n12,
  n13,
  n23,
  n123,
  labels = c("A", "B", "C"),
  title = "Overlap",
  quantities = c("counts", "percent"),
  png_file = NULL,
  height = 5,
  width = 5,
  dpi = 150,
  colors = c("#FF6D60", "#176B87", "#FF8400"),
  ...
)
}
\arguments{
\item{n1}{Input integer. The number of components only in the first group,
must be positive integer.}

\item{n2}{Input integer. The number of components only in the second group,
must be positive integer.}

\item{n3}{Input integer. The number of components only in the third group,
must be positive integer.}

\item{n12}{Input integer. The number of components overlapped between
1 and 2, should be positive integer or zero.}

\item{n13}{Input integer. The number of components overlapped between
1 and 3, should be positive integer or zero.}

\item{n23}{Input integer. The number of components overlapped between
2 and 3, should be positive integer or zero.}

\item{n123}{Input integer. The number of components overlapped between
all three groups, should be positive integer or zero.}

\item{labels}{Input vector. The name of the groups, length == 3}

\item{title}{Input vector. The title of the plot, default: Overlap}

\item{quantities}{character. could be \code{counts} or \code{percent}}

\item{png_file}{Input vector. NULL or path to the output png file, default
: NULL}

\item{height}{Input vector. The height of output file in inches, default: 5}

\item{width}{Input vector. The width of output file in inches, default: 5}

\item{dpi}{Input vector. The resolution of output file, default: 150}

\item{colors}{character. The colors for each group,}

\item{...}{A series of named character vectors. parameters for \code{\link[base:plot]{base::plot()}}
function}
}
\value{
A plot
}
\description{
generate overlap between three groups
\code{overlap3} create venn plot for three groups
}
\examples{
overlap2(3, 8, 1)
overlap2(12, 8, 4, name1 = "group1", name2 = "group2")

}
\seealso{
\link[VennDiagram:venn.diagram]{VennDiagram::venn.diagram} which doing the same job
}
