% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{overlap4}
\alias{overlap4}
\title{recommend ggupset and UpSetR for more than 3 groups overlap plot}
\usage{
overlap4(...)
}
\arguments{
\item{...}{Input vector. arguments}
}
\description{
recommend ggupset and UpSetR for more than 3 groups overlap plot
}
