% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_plot}
\alias{save_plot}
\title{save plot as pdf}
\usage{
save_plot(
  x,
  file,
  overwrite = FALSE,
  save_rds = FALSE,
  call = rlang::caller_env(),
  ...
)
}
\arguments{
\item{x}{graph vector. Could be ggplot or grob}

\item{file}{character. Path to the pdf file}

\item{overwrite}{bool. Overwrite exists files, default: FALSE}

\item{save_rds}{bool. Save plot object as RDS, default: FALSE}

\item{call}{vector. for error report, see \code{\link[rlang:stack]{rlang::caller_env()}}}

\item{...}{arguments for \link[export:graph2vector]{export::graph2pdf} function, such as \code{width},
\code{height}, \code{font}, \code{bg}, ...}
}
\value{
NULL. return NULL
}
\description{
save plot as pdf
}
\examples{
\dontrun{
p <- plot(1:10)
save_plot(p, "demo.png")
save_plot(p, "demo.pdf", width = 4, height = 3, bg = "white")
}

}
